Move transactions, wallets, categories, and budgets into /dashboard
