+++ src/types/index.ts
@@ -1,6 +1,6 @@
 export * from "./transaction";
-import { getUserCategories } from "@/app/categories/category.db";
-import { getUserWalletsWithBalance } from "@/app/wallets/wallet.actions";
+import { getUserCategories } from "@/app/dashboard/categories/category.db";
+import { getUserWalletsWithBalance } from "@/app/dashboard/wallets/wallet.actions";
 
 export type Category = Awaited<ReturnType<typeof getUserCategories>>["data"][0];
 export type WalletWithBalance = Awaited<
